+++ packages/design-system/package.json
@@ -43,8 +43,10 @@
     "./components/Label": "./dist/components/Form/Label.js",
     "./components/Separator": "./dist/components/Separator/index.js",
     "./molecules/AsideBrandLayout": "./dist/molecules/AsideBrandLayout/index.js",
-    "./modules/auth/SignInForm": "./dist//modules/auth/SignInForm/index.js",
-    "./modules/auth/serverApp": "./dist//modules/auth/serverApp.js",
+    "./modules/auth/SignInForm": "./dist/modules/auth/SignInForm/index.js",
+    "./modules/auth/serverApp": "./dist/modules/auth/serverApp.js",
+    "./modules/auth/serviceWorker": "./dist/modules/auth/serviceWorker.js",
+    "./modules/auth/hooks": "./dist/modules/auth/hooks.js",
     "./forms/Field": "./dist/forms/Field/index.js",
     "./forms/useForm": "./dist/forms/useForm/index.js",
     "./main.css": "./dist/main.css"
